Move CGO files to main package directory.
Per https://github.com/tools/godep/issues/422, godep won't find cgo files in
other directories, so vendoring won't work without this fix.
