+++ hat-trie.c
