+++ hat-trie.h
