--- hat-trie/src/Makefile.am
@@ -1,11 +1,0 @@
-
-lib_LTLIBRARIES = libhat-trie.la
-
-libhat_trie_la_SOURCES = common.h \
-                         ahtable.h        ahtable.c \
-                         hat-trie.h       hat-trie.c \
-                         misc.h           misc.c \
-                         murmurhash3.h    murmurhash3.c
-
-pkginclude_HEADERS = hat-trie.h ahtable.h common.h pstdint.h
-
