+++ misc.c
