+++ misc.h
