+++ murmurhash3.c
